webservice/requirements.txt
aniso8601==7.0.0
Click==7.0
Flask==1.1.1
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pyblake2==1.1.2
python-bitcoinrpc==1.0
pytz==2019.1
six==1.12.0
SQLAlchemy==1.3.5
Werkzeug==0.15.5
git+https://github.com/haobtc/pycoin.git@hardfork_solutions 
